// src/modules/api/model/taskDtoStatus.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Ambeaver API
 * OpenAPI spec version: 0.0.1
 */

export type TaskDtoStatus = (typeof TaskDtoStatus)[keyof typeof TaskDtoStatus];

export const TaskDtoStatus = {
  FINISHED: "FINISHED",
  READY_FOR_CLAIM: "READY_FOR_CLAIM",
  NOT_STARTED: "NOT_STARTED",
} as const;